+++ pkg/telemetry/data.go
@@ -2,7 +2,6 @@ package telemetry
 
 import (
 	"encoding/json"
-	"github.com/SAP/jenkins-library/pkg/log"
 	"net/url"
 )
 
@@ -95,23 +94,4 @@ func (d *Data) toPayloadString() string {
 	}
 
 	return parameters.Encode()
-}
-
-// PipelineTelemetry object to store pipeline related telemetry information
-type PipelineTelemetry struct {
-	CorrelationId       string             `json:"CorrelationId"`       // CorrelationId
-	Duration            string             `json:"PipelineDuration"`    // Duration of the pipeline in milliseconds
-	Orchestrator        string             `json:"Orchestrator"`        // Orchestrator, e.g. Jenkins or Azure
-	OrchestratorVersion string             `json:"OrchestratorVersion"` // OrchestratorVersion
-	PipelineStartTime   string             `json:"PipelineStartTime"`   // PipelineStartTime Pipeline start time
-	Errors              []log.ErrorDetails `json:"ErrorDetails"`        // Errors Error Codes found in errorsJson
-	BuildId             string             `json:"BuildId"`             // BuildId of the pipeline run
-	JobName             string             `json:"JobName"`
-	PipelineStatus      string             `json:"PipelineStatus"`
-	GitInstance         string             `json:"GitInstance"`
-	JobURL              string             `json:"JobURL"`
-	CommitHash          string             `json:"CommitHash"`
-	Branch              string             `json:"Branch"`
-	GitOwner            string             `json:"GitOwner"`
-	GitRepository       string             `json:"GitRepository"`
 }
